small touches on gpu code

+++ RecoLocalCalo/HcalRecAlgos/interface/MahiFit_gpu.h
@@ -3,7 +3,7 @@
 
 #include "DataFormats/HcalDetId/interface/HcalSubdetector.h"
 #include "DataFormats/HcalDetId/interface/HcalDetId.h"
-#include "RecoLocalCalo/HcalRecAlgos/interface/EigenMatrixTypes.h"
+#include "RecoLocalCalo/HcalRecAlgos/interface/EigenMatrixTypes_gpu.h"
 #include "DataFormats/HcalRecHit/interface/HBHEChannelInfo.h"
 
 #include "RecoLocalCalo/HcalRecAlgos/interface/PulseShapeFunctor_gpu.h"
@@ -132,7 +132,7 @@ class MahiFit
   __device__
   void updateCov() const;
   __device__
-  void updatePulseShape(double itQ, FullSampleVector &pulseShape, 
+  void updatePulseShape(float itQ, FullSampleVector &pulseShape,
 			FullSampleVector &pulseDeriv,
 			FullSampleMatrix &pulseCov) const;
